Implemented Redux
Set Up Redux Functionality

Added Redux to handle our app's state, and better handle deleting an
event.

Renamed ShoppingList to InventoryList

Since this is an inventory app, it makes sense to have a view be called
InventoryList rather than ShoppingList. Therefore, ShoppingList was
renamed.

Finished Implementing Redux

Redux now handles adding items and deleting items. Now that the
functionality is there, it can be made use of in the future.