Implemented Sale Functionality
Implemented Search By Barcode

The Sale View now acceps a barcode entry in an input box, it can then
search the database by the value entered, and return the result in a
modal.

Added Backend Functionality to Edit Item

We can now edit an item in the front end and have a process be done in
mongodb to udpate the item's corresponding fields.

Added Handler to Stop Invalid Quantity Entry

Now users can only enter a number between 0 and the item's maximum
price in the sale modal.

Made Sale Modal Prettier

Commented Line out in Item.js

This was done as there was a bug in rendering the item with searching.

Updated Sale Modal Contents

The sale modal now displays price of item, and the updated quantity
after the sale.

Added Functionality to Update Inventory Item

Using a post request, we can look for an item to update if an id was
sent with the request.

Added Ability to Log Purchases

iEcho can now search for an item by its name or barcode, and the user
can specify how much of an item they wish to buy in a sale modal. This
purchase is then logged in mongoDB, and shown in the sale view page
of the application.